♻️ refactor : api 함수 위치변경으로 인한 경로 재설정
